feat: Helm templates in provided secret names.
Added support for helm templates in the following entries:
* `.existingSecret`
* `.sentinel.existingSecret`
* `.haproxy.tls.secretName`
* `.tls.secretName`
* `.restore.existingSecret`

We are using this great `redis-ha` chart as a sub-chart and we provide TLS secret and credentials via secrets (existing).
We would like not to hardcode these name in values.yaml (i.e., tweak them for each installation)
  but instead use some reasonable default like `{{ .Release.Name }}-redis-creds`

Replacing:
`.Values.existingSecret`  by ` tpl (.Values.existingSecret | default "" ) .`

Note that  `| default ""` is required because the `existingSecret` is commented in the chart default values.yaml and thus undefined for helm templating --> without the default values the rendering would result into errors.

Signed-off-by: Pavel Jancik <pavel.jancik@themama.ai>