redis-ha: make readiness and liveness probes optional, add startup probe
Allow disabling the readiness and liveness probes on both Redis and
Sentinel. This can be useful if issues are encountered with the probes.
Also add a startup probe which can better handle the sometimes lengthy
process of loading a large database and subsequently refreshing the
local copy from the primary / master server.

Signed-off-by: Chris Boot <bootc@bootc.net>