redis-ha: fix busybox-ash-ism in redis_liveness.sh
The ${response:0:7} syntax is non-POSIX and not understood by Debian's
dash shell, so if a Debian-based Redis image is used the liveness check
breaks on slaves. A case statement makes this check more legible anyway,
in my opinion, and works in all the POSIX-compatible shells (busybox
ash, dash, and bash).

Signed-off-by: Chris Boot <bootc@bootc.net>